+++ package.json
@@ -12,9 +12,9 @@
     "lint": "eslint --ext .ts --ext .tsx src/",
     "typecheck": "tsc --noEmit",
     "clean": "rm -rf dist/",
-    "start": "node dist/index.js",
+    "start": "node dist/src/main.js",
     "build": "npm run clean && tsc",
-    "dev": "rm -rf testingname/ testing-name/; npm run build; node dist/src/index.js testing-name",
+    "dev": "rm -rf testingname/ testing-name/; npm run build; node dist/src/main.js testing-name",
     "test": "mocha",
     "test:full": "npm run lint; npm run typecheck; mocha",
     "prepublishOnly": "npm run lint && npm run test",
